Use less precise intent filter for listening for NFC YubiKeys
It seems that the previous method doesn't work properly on some devices or not with every YubiKey. This new approach doesn't strictly verify that the NFC tag is actually a YubiKey and will allow to use any ISO 14443-4 tag. Let's just hope that the user won't confuse his credit card for a YubiKey.

Fixes #2